refactor(loader): remove code repetitions
