+++ include/tools/loader/loader_linux.h
@@ -8,32 +8,13 @@
 #include <string>
 #include <string_view>
 
+#include "loader_shared.h"
+
 // keep in sync with get_module_file_name()
 #define LOADER_LINUX_GET_MODULE_FILE_NAME(name) "lib" name ".so"
 
 namespace loader::linux {
     using module_handle_t = void*;
-
-    class LoadModuleError {
-    public:
-        /// @return Lifetime bound to this @ref LoadModuleError
-        [[nodiscard]] auto as_string() const -> std::string_view {
-            return this->m_errorMessage;
-        }
-
-        static auto from_string(std::string_view str) {
-            return LoadModuleError{str};
-        }
-
-    private:
-        // we can't use allocating C++ functions (std::string) in here, see doc
-        // of LoadModuleError in loader.h.
-        char m_errorMessage[512]{};
-
-        explicit LoadModuleError(std::string_view str) {
-            std::strncpy(this->m_errorMessage, str.data(), std::min(std::size(m_errorMessage), std::size(str)));
-        }
-    };
 
     // keep in sync with LOADER_LINUX_GET_MODULE_FILE_NAME
     [[nodiscard]] inline auto get_module_file_name(std::string name) -> std::string {
